- Remove unused prop from EntityBadgeIcons.tsx - Typechecking for EntityBadgeIcons tests - Move table cell components to individual files
