+++ packages/synapse-react-client/src/components/EntityFinder/details/view/table/DateCell.tsx
@@ -1,0 +1,6 @@
+import dayjs from 'dayjs'
+import { formatDate } from '../../../../../utils/functions/DateFormatter'
+
+export function DateCell({ cellData }: { cellData?: string }) {
+  return <>{(cellData && formatDate(dayjs(cellData))) ?? <></>}</>
+}
