+++ packages/synapse-react-client/src/components/EntityFinder/details/view/table/EntityFinderCheckboxCell.tsx
@@ -1,0 +1,23 @@
+import { Checkbox } from '@mui/material'
+import { CellContext } from '@tanstack/react-table'
+
+export function EntityFinderCheckboxCell<
+  T extends { entityId: string; isSelected: boolean; isDisabled?: boolean },
+>(props: CellContext<T, any>) {
+  const { row } = props
+  const { entityId, isSelected, isDisabled } = row.original
+
+  if (isDisabled) {
+    return null
+  }
+
+  return (
+    <Checkbox
+      inputProps={{ 'aria-label': `Select ${entityId}` }}
+      checked={isSelected}
+      onChange={() => {
+        // no-op
+      }}
+    />
+  )
+}
