+++ packages/synapse-react-client/src/components/EntityFinder/details/view/table/EntityTypeCell.tsx
@@ -1,0 +1,16 @@
+import { EntityType } from '@sage-bionetworks/synapse-types'
+import { CellContext } from '@tanstack/react-table'
+import { EntityTypeIcon } from '../../../../EntityIcon'
+import { EntityFinderTableViewRowData } from '../DetailsView'
+
+export function EntityTypeCell(
+  props: CellContext<EntityFinderTableViewRowData, EntityType>,
+) {
+  const { getValue } = props
+  return (
+    <EntityTypeIcon
+      className="EntityFinderTableCellEntityIcon"
+      type={getValue()}
+    />
+  )
+}
