+++ packages/synapse-react-client/src/components/EntityFinder/details/view/table/ModifiedByCell.tsx
@@ -1,0 +1,26 @@
+import { Skeleton } from '@mui/material'
+import { CellContext } from '@tanstack/react-table'
+import useGetEntityBundle from '../../../../../synapse-queries/entity/useEntityBundle'
+import { EntityIdAndVersionNumber } from './TableCellTypes'
+import { UserBadgeCell } from './UserBadgeCell'
+
+/**
+ * Renders 'modifiedBy' from the entity bundle.
+ * @param props
+ * @returns
+ */
+export function ModifiedByCell<T extends EntityIdAndVersionNumber>(
+  props: CellContext<T, any>,
+) {
+  const { row } = props
+  const { data: bundle, isLoading } = useGetEntityBundle(
+    row.original.entityId,
+    row.original.versionNumber,
+  )
+
+  if (isLoading) {
+    return <Skeleton width={200} />
+  }
+
+  return <UserBadgeCell {...props} cellData={bundle?.entity.modifiedBy} />
+}
