+++ packages/synapse-react-client/src/components/EntityFinder/details/view/table/UserBadgeCell.tsx
@@ -1,0 +1,9 @@
+import { UserBadge } from '../../../../UserCard/UserBadge'
+
+export function UserBadgeCell({ cellData }: { cellData?: string }) {
+  return cellData ? (
+    <UserBadge userId={cellData} openLinkInNewTab={true} />
+  ) : (
+    <></>
+  )
+}
